#2056: Move fixtures to root conftest
